拆分数据库表，并加上约束条件。 添加menu表， 不同身份用户登录后显示不同场景的校验 前端代码全部用vscode开发

+++ src/main/java/cn/edu/tjut/stud/domain/Course.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class Course
 {
     @TableId
-    private Long cno;
+    private int cno;
     private String cname;
     private String type;
     private Double score;
